+++ src/test/scala/dxWDL/compiler/InputFileTest.scala
@@ -119,9 +119,42 @@ class InputFileTest extends FlatSpec with Matchers {
     }
 
 
-    it should "support struct inputs" taggedAs(EdgeTest) in {
+    it should "support struct inputs" in {
         val wdlCode = pathFromBasename("struct", "Person.wdl")
         val inputs = pathFromBasename("struct", "Person_input.json")
+
+        val retval = Main.compile(
+            List(wdlCode.toString, "-inputs", inputs.toString)
+                ++ cFlags
+        )
+        retval shouldBe a [Main.SuccessfulTerminationIR]
+    }
+
+    it should "support array of pairs" in {
+        val wdlCode = pathFromBasename("input_file", "echo_pairs.wdl")
+        val inputs = pathFromBasename("input_file", "echo_pairs.json")
+
+        val retval = Main.compile(
+            List(wdlCode.toString, "-inputs", inputs.toString)
+                ++ cFlags
+        )
+        retval shouldBe a [Main.SuccessfulTerminationIR]
+    }
+
+    it should "array of structs" in {
+        val wdlCode = pathFromBasename("struct", "array_of_structs.wdl")
+        val inputs = pathFromBasename("struct", "array_of_structs_input.json")
+
+        val retval = Main.compile(
+            List(wdlCode.toString, "-inputs", inputs.toString)
+                ++ cFlags
+        )
+        retval shouldBe a [Main.SuccessfulTerminationIR]
+    }
+
+    ignore should "override default values in input file" taggedAs(EdgeTest) in {
+        val wdlCode = pathFromBasename("input_file", "override.wdl")
+        val inputs = pathFromBasename("input_file", "override_input.json")
 
         val retval = Main.compile(
             List(wdlCode.toString, "-inputs", inputs.toString)